pagination added

+++ app/Http/Controllers/ProfileController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use Illuminate\Http\Request;
 use Auth;
 use App\User;
+use App\Note;
 use Image;
 use Illuminate\Support\Facades\Storage;
 class ProfileController extends Controller
@@ -64,7 +65,9 @@ class ProfileController extends Controller
 
             $user_id = Auth::user()->id;
             $user = User::find(Auth::user()->id);
+            if($user->avatar != 'default.jpg'){
             Storage::delete('public/user_avatar/'.$user->avatar);
+            }
             $user->avatar =  $avatarNameToStore;
 
             // $path = $request->file('avatar')->resize(150,150)->storeAs('public/user_avatar', $avatarNameToStore);
@@ -83,9 +86,10 @@ class ProfileController extends Controller
     public function show($id)
     {
         $user = User::where('id',$id)->get();
+        $notes = Note::where('user_id',$id)->paginate(5);
         // dd($user);
         $user_id = $id;
-        return view('profile.show')->with('user_id',  $user_id )->with('user',$user);
+        return view('profile.show')->with('user_id',  $user_id )->with('user',$user)->with('notes',$notes);
     }
 
     /**
